// shared/schemas.d.ts
export declare const notarySchema: z.ZodObject<{
    id: z.ZodOptional<z.ZodString>;
    full_name: z.ZodString;
    cin: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    tax_id: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    appointment_number: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    start_date: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    office_location: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    region: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    phone: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    email: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    photo_url: z.ZodNullable<z.ZodOptional<z.ZodString>>;
    description: z.ZodNullable<z.ZodOptional<z.ZodString>>;
}, "strip", z.ZodTypeAny, {
    id?: string;
    full_name?: string;
    cin?: string;
    tax_id?: string;
    appointment_number?: string;
    start_date?: string;
    office_location?: string;
    region?: string;
    phone?: string;
    email?: string;
    photo_url?: string;
    description?: string;
}, {
    id?: string;
    full_name?: string;
    cin?: string;
    tax_id?: string;
    appointment_number?: string;
    start_date?: string;
    office_location?: string;
    region?: string;
    phone?: string;
    email?: string;
    photo_url?: string;
    description?: string;
}>;
// ...
export type Notary = z.infer<typeof notarySchema>;